% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brts2phylo.R
\name{brts2phylo}
\alias{brts2phylo}
\title{Converts brts to phylo}
\usage{
brts2phylo(times, root = FALSE, tip.label = NULL)
}
\arguments{
\item{times}{brts times}

\item{root}{bool for rooting}

\item{tip.label}{bool for labels}
}
\value{
A phylo
}
\description{
A copy of function from DDD, to debug on GHA
}
