% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_phylo.R
\name{convert_to_phylo}
\alias{convert_to_phylo}
\title{Converts list of branching time vectors to phylo objects}
\usage{
convert_to_phylo(brts)
}
\arguments{
\item{brts}{List with vector of branching times as returned by}
}
\value{
List with phylo objects.
}
\description{
Converts list of branching time vectors to phylo objects
}
