% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{Default parameter documentation}
\usage{
default_params_doc(totaltime, sim_pars, n_mainland_spec, replicates)
}
\arguments{
\item{totaltime}{Positive numeric. Specifies how the age of the island to be
simulated by \code{\link[DAISIE:DAISIE_sim]{DAISIE::DAISIE_sim_constant_rate()}}.}

\item{sim_pars}{Numeric vector with DAISIE simulation parameters, as in
\code{\link[DAISIE:DAISIE_sim]{DAISIE::DAISIE_sim_constant_rate()}}:
\enumerate{
\item Cladogenesis rate (\eqn{\lambda^c}).
\item Extinction rate \eqn{\mu}.
\item Island-wide carrying capacity (K').
\item Immigration rate \eqn{\gamma}.
\item Anagenesis rate \eqn{\lambda^a}.
}}

\item{n_mainland_spec}{Positive Integer with total number of species in the
mainland pool, which can colonize the island.}

\item{replicates}{Positive integer with the total number of replicates of the
same island that should be simulated.}
}
\value{
Nothing
}
\description{
This function does not do anything. It is meant to hold argument
documentation, which is then inherited in all other functions.
}
\note{
Should be marked as \verb{@noRd}, but this can't be done or inheriting
parameters would not be possible.
}
\author{
Pedro Neves
}
