% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_first_lineage.R
\name{extract_first_lineage_brts}
\alias{extract_first_lineage_brts}
\title{Extract first lineagebranching times vector}
\usage{
extract_first_lineage_brts(iw_sims)
}
\arguments{
\item{iw_sims}{DAISIE object (list) as generated by \code{\link[=run_iw_sims]{run_iw_sims()}}.}
}
\value{
List with a level per replicate, each containing a numeric vector
with branching times of first lineage of the simulation.
}
\description{
Extract first lineagebranching times vector
}
\examples{
out <- phylometricsims:::run_iw_sims(
  totaltime = 4,
  sim_pars = c(1, 2, 3, 4, 5),
  n_mainland_spec = 1000,
  n_replicates = 2,
  seed_start = 1
)

brts_list <- phylometricsims:::extract_first_lineage_brts(out)

}
\author{
Pedro Neves
}
