% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pipeline.R
\name{run_pipeline}
\alias{run_pipeline}
\title{Run the pipeline for a single parameter set}
\usage{
run_pipeline(param_set, n_replicates = 1, seed_start = 1)
}
\arguments{
\item{param_set}{Positive integer. Corresponds to one line of the param_space
data frame.}

\item{n_replicates}{Positive integer with the total number of replicates of
the same island that should be simulated.}

\item{seed_start}{Integer specifying the seed of replicate 1. Each replicate
a given param set will be simulated using seed 1:n_replicates.}
}
\value{
Save output to output and trees folder. Generate .tree file if >
20 tips are created.
}
\description{
Run the pipeline for a single parameter set
}
\examples{
run_pipeline(param_set = 1)
}
