% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_seldig_experiment.R
\name{run_seldig_experiment}
\alias{run_seldig_experiment}
\title{Generate phylogenies for sELDIG study}
\usage{
run_seldig_experiment(n_param_sets_to_run, n_replicates = 1)
}
\arguments{
\item{n_param_sets_to_run}{Integer with number of parameter sets (i.e.,
columns in the parameter space file) should be simulated.}

\item{n_replicates}{Positive integer with the total number of replicates of
the same island that should be simulated.}
}
\value{
Saves output object to output and trees folder. Generate .tree file if >
20 tips are created.
Output object is a list containing the following elements:
\itemize{
\item DAISIE IW simulation objects
\item Corresponding vector of branching times
\item Corresponding branching times converted to phylo objects
}
}
\description{
Run DAISIE IW simulations and save .tre files of the first lineage on the
island if the number of tips at the present exceeds 20. Save intermediate
output to /output folder, see Return section for details.
}
\examples{
run_seldig_experiment(n_param_sets_to_run = 2)
}
\seealso{
\code{\link[=run_pipeline]{run_pipeline()}}
}
\author{
Pedro Neves
}
