% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_phylo.R
\name{convert_to_phylo}
\alias{convert_to_phylo}
\title{Converts list of branching time vectors to phylo objects}
\usage{
convert_to_phylo(brts, seed_start)
}
\arguments{
\item{brts}{List with vector of branching times as returned by}

\item{seed_start}{Integer specifying the seed of replicate 1. Each replicate
a given param set will be simulated using seed \code{1:n_replicates}.}
}
\value{
List with phylo objects.
}
\description{
Converts list of branching time vectors to phylo objects
}
