% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{Default parameter documentation}
\usage{
default_params_doc(
  totaltime,
  sim_pars,
  n_mainland_spec,
  n_replicates,
  seed_start,
  iw_sims,
  param_set,
  brts,
  n_param_sets_to_run
)
}
\arguments{
\item{totaltime}{Positive numeric. Specifies how the age of the island to be
simulated by \code{\link[DAISIE:DAISIE_sim]{DAISIE::DAISIE_sim_constant_rate()}}.}

\item{sim_pars}{Numeric vector with DAISIE simulation parameters, as used by
\code{\link[DAISIE:DAISIE_sim]{DAISIE::DAISIE_sim_constant_rate()}}:
\enumerate{
\item simID, an integer with the ID of the parameters. Not used in the
simulation.
\item Cladogenesis rate (\eqn{\lambda^c}).
\item Extinction rate \eqn{\mu}.
\item Island-wide carrying capacity (K').
\item Immigration rate \eqn{\gamma}.
\item Anagenesis rate \eqn{\lambda^a}.
\item Totaltime: the island age.
}}

\item{n_mainland_spec}{Positive integer with total number of species in the
mainland pool, which can colonize the island.}

\item{n_replicates}{Positive integer with the total number of replicates of
the same island that should be simulated.}

\item{seed_start}{Integer specifying the seed of replicate 1. Each replicate
a given param set will be simulated using seed \code{1:n_replicates}.}

\item{iw_sims}{DAISIE object (list) as generated by \code{\link[=run_iw_sims]{run_iw_sims()}}.}

\item{param_set}{Positive integer. Corresponds to one line of the param_space
data frame.}

\item{brts}{List with vector of branching times as returned by}

\item{n_param_sets_to_run}{Integer with number of parameter sets (i.e.,
columns in the parameter space file) should be simulated.}
}
\value{
Nothing
}
\description{
This function does not do anything. It is meant to hold argument
documentation, which is then inherited in all other functions.
}
\note{
Should be marked as \verb{@noRd}, but this can't be done or inheriting
parameters would not be possible.
}
\author{
Pedro Neves
}
