% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pipeline.R
\name{run_pipeline}
\alias{run_pipeline}
\title{Run the pipeline for a single parameter set}
\usage{
run_pipeline(n_param_sets_to_run, save_to_file = TRUE, test_seed = FALSE)
}
\arguments{
\item{n_param_sets_to_run}{Integer with number of parameter sets (i.e.,
columns in the parameter space file) should be simulated.}

\item{save_to_file}{Boolean stating if output of pipeline should be saved
as an \code{.rds} file in \verb{./output/}}

\item{test_seed}{A boolean, which sets the seed to 1. Defaults to \code{FALSE},
should only be set to \code{TRUE} for testing.}
}
\value{
Saves output object to output and trees folder. Generate .tree file if >
20 tips are created.
Output object is a list containing the following elements:
\itemize{
\item DAISIE IW simulation objects
\item Corresponding vector of branching times
\item Corresponding branching times converted to phylo objects
}
}
\description{
Runs the full pipeline one parameter set (i.e., one row of the
param_space.csv file). DAISIE simulations are generated, the first lineage,
if any, is extracted. Associated branching times are converted to a phylo
object which is then saved as a .tre file to the /trees folder, if there are
more than 20 tips. Intermediate objects are saved as they are generated to
output folder, as a list. See details in the Return section.
}
\author{
Pedro Neves
}
