% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pipeline.R
\name{run_pipeline}
\alias{run_pipeline}
\title{Run the pipeline for a single parameter set}
\usage{
run_pipeline(param_set, n_replicates = 1, seed_start = 1, save_to_file = TRUE)
}
\arguments{
\item{param_set}{Positive integer. Corresponds to one line of the param_space
data frame.}

\item{n_replicates}{Positive integer with the total number of replicates of
the same island that should be simulated.}

\item{seed_start}{Integer specifying the seed of replicate 1. Each replicate
a given param set will be simulated using seed \code{1:n_replicates}.}
}
\value{
Saves output object to output and trees folder. Generate .tree file if >
20 tips are created.
Output object is a list containing the following elements:
\itemize{
\item DAISIE IW simulation objects
\item Corresponding vector of branching times
\item Corresponding branching times converted to phylo objects
}
}
\description{
Runs the full pipeline one parameter set (i.e., one row of the
param_space.csv file). DAISIE simulations are generated, the first lineage,
if any, is extracted. Associated branching times are converted to a phylo
object which is then saved as a .tre file to the /trees folder, if there are
more than 20 tips. Intermediate objects are saved as they are generated to
output folder, as a list. See details in the Return section.
}
\author{
Pedro Neves
}
